pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "flagship-lite"
version = "0.1.0"
description = "Feature flags that live in your repo. YAML-based, evaluated locally, no SaaS needed."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.9"
authors = [{name = "flagship-lite contributors"}]
keywords = ["feature-flags", "feature-toggle", "yaml", "rollout", "targeting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "click>=8.0",
    "pyyaml>=6.0",
    "fastapi>=0.110",
    "uvicorn>=0.27",
]

[project.scripts]
flagship-lite = "flagship_lite.cli:cli"

[project.urls]
Homepage = "https://github.com/user/flagship-lite"
Issues = "https://github.com/user/flagship-lite/issues"
